field_services_app/requirements.txt
# Core dependencies
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
Pillow==9.5.0
celery==5.2.7
redis==4.5.4

# Authentication
djangorestframework-simplejwt==5.2.2
django-allauth==0.54.0

# Project management
django-mptt==0.14.0  # For tree structures (e.g., task hierarchies)

# Documentation
drf-yasg==1.21.5  # For API documentation

# WhatsApp integration
twilio==8.0.0  # For WhatsApp Business API via Twilio

# Geo & routing
geopy==2.3.0
folium==0.14.0  # For map visualizations

# Utilities
python-dotenv==1.0.0
django-filter==23.1
django-crispy-forms==2.0
django-storages==1.13.2
boto3==1.26.115  # For AWS S3 integration

# Development
black==23.3.0
flake8==6.0.0
pytest==7.3.1
pytest-django==4.5.2
faker==18.3.1  # For generating test data

# Deployment
gunicorn==20.1.0
whitenoise==6.4.0  # For serving static files
